Add a project. Add ffi library files. Add config.h Small fix to suppress warnings.

+++ src/struct.c
@@ -371,7 +371,7 @@ static int b_unpack (lua_State *L) {
         if (size == 0) {
           if (!lua_isnumber(L, -1))
             luaL_error(L, "format `c0' needs a previous size");
-          size = lua_tonumber(L, -1);
+          size = (size_t)lua_tonumber(L, -1);
           lua_pop(L, 1);
           luaL_argcheck(L, pos+size <= ld, 2, "data string too short");
         }
